django/pyproject.toml
[tool.poetry]
name = "eltexsoft-url-shortener-task"
version = "0.1.0"
description = ""
authors = ["Yevhen Shymotyuk <yevhen.shymotyuk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "*"
python-dotenv = "*"
python-decouple = "*"
psycopg2-binary = "*"
coverage = "*"
django-extensions = "*"
pillow = "*"
gunicorn = "*"
djangorestframework = "^3.10"
django-rest-swagger = "^2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
